@@ -18,51 +18,46 @@
     "synth": "npx projen synth",
     "synth:silent": "npx projen synth:silent",
     "test": "npx projen test",
-    "test:update": "npx projen test:update",
     "test:watch": "npx projen test:watch",
     "watch": "npx projen watch",
     "projen": "npx projen"
   },
   "devDependencies": {
-    "@aws-cdk/assert": "1.152.0",
-    "@aws-cdk/assertions": "1.152.0",
-    "@types/jest": "^27.4.0",
-    "@types/node": "^12",
+    "@types/fs-extra": "^9.0.13",
+    "@types/jest": "^29.2.4",
+    "@types/node": "^16",
+    "@types/source-map-support": "^0.5.6",
     "@typescript-eslint/eslint-plugin": "^5",
     "@typescript-eslint/parser": "^5",
-    "aws-cdk": "1.152.0",
-    "esbuild": "^0.14.14",
+    "aws-cdk": "2.56.1",
+    "esbuild": "^0.16.10",
     "eslint": "^8",
     "eslint-import-resolver-node": "^0.3.6",
-    "eslint-import-resolver-typescript": "^2.5.0",
-    "eslint-plugin-import": "^2.25.4",
-    "jest": "^27.4.7",
+    "eslint-import-resolver-typescript": "^3.5.2",
+    "eslint-plugin-import": "^2.26.0",
+    "fs-extra": "*",
+    "jest": "^29.3.1",
     "jest-junit": "^13",
     "json-schema": "^0.4.0",
-    "projen": "^0.54.22",
-    "ts-jest": "^27.1.3",
-    "ts-node": "^9",
-    "typescript": "^4.5.5"
+    "projen": "^0.65.72",
+    "ts-jest": "^29.0.3",
+    "ts-node": "^10.9.1",
+    "typescript": "^4.9.4"
   },
   "dependencies": {
-    "@aws-cdk/aws-cloudwatch": "1.152.0",
-    "@aws-cdk/aws-dynamodb": "1.152.0",
-    "@aws-cdk/aws-events": "1.152.0",
-    "@aws-cdk/aws-events-targets": "1.152.0",
-    "@aws-cdk/aws-iam": "1.152.0",
-    "@aws-cdk/aws-lambda": "1.152.0",
-    "@aws-cdk/aws-lambda-nodejs": "1.152.0",
-    "@aws-cdk/aws-secretsmanager": "1.152.0",
-    "@aws-cdk/aws-ssm": "1.152.0",
-    "@aws-cdk/core": "1.152.0",
-    "constructs": "^3.2.27"
+    "aws-cdk-lib": "2.56.1",
+    "constructs": "^10.0.5",
+    "source-map-support": "^0.5.13"
+  },
+  "engines": {
+    "node": ">= 16.18.1"
   },
   "license": "Apache-2.0",
   "version": "0.0.0",
   "jest": {
     "testMatch": [
       "<rootDir>/src/**/__tests__/**/*.ts?(x)",
-      "<rootDir>/(test|src)/**/?(*.)+(spec|test).ts?(x)"
+      "<rootDir>/(test|src)/**/*(*.)@(spec|test).ts?(x)"
     ],
     "clearMocks": true,
     "collectCoverage": true,
