@@ -1,8 +1,9 @@
-import { Table } from '@aws-cdk/aws-dynamodb';
-import { IFunction, Runtime } from '@aws-cdk/aws-lambda';
-import { NodejsFunction } from '@aws-cdk/aws-lambda-nodejs';
-import { Secret } from '@aws-cdk/aws-secretsmanager';
-import { Construct, Duration, Stack, StackProps } from '@aws-cdk/core';
+import { Duration, Stack, StackProps } from 'aws-cdk-lib';
+import { Table } from 'aws-cdk-lib/aws-dynamodb';
+import { IFunction, Runtime } from 'aws-cdk-lib/aws-lambda';
+import { NodejsFunction } from 'aws-cdk-lib/aws-lambda-nodejs';
+import { Secret } from 'aws-cdk-lib/aws-secretsmanager';
+import { Construct } from 'constructs';
 
 export interface ApiStackProps extends StackProps {
   stage: string;
